LPS-65351 Introduce extensible factory pattern for constructing SAQContext which enables SAQ to be used in other contexts than AccessControlPolicy
