+++ modules/apps/foundation/portal-security/portal-security-service-access-quota-service/src/main/java/com/liferay/portal/security/service/access/quota/internal/impl/UserIdSAQMetricProvider.java
@@ -15,7 +15,7 @@
 package com.liferay.portal.security.service.access.quota.internal.impl;
 
 import com.liferay.portal.kernel.security.auth.AccessControlContext;
-import com.liferay.portal.security.service.access.quota.SAQMetricProvider;
+import com.liferay.portal.security.service.access.quota.AccessControlPolicySAQMetricProvider;
 
 import java.lang.reflect.Method;
 
@@ -24,8 +24,9 @@ import org.osgi.service.component.annotations.Component;
 /**
  * @author Stian Sigvartsen
  */
-@Component(service = SAQMetricProvider.class)
-public class UserIdSAQMetricProvider implements SAQMetricProvider {
+@Component(service = AccessControlPolicySAQMetricProvider.class)
+public class UserIdSAQMetricProvider
+	implements AccessControlPolicySAQMetricProvider {
 
 	@Override
 	public String getMetricName() {
